+++ example/lib/app_routes.dart
@@ -1,6 +1,8 @@
+import 'package:flutter/material.dart';
 import 'package:flutter_link_nav/flutter_link_nav.dart';
 
 import 'detail_screen.dart';
+import 'main.dart';
 import 'main_screen.dart';
 
 class ExampleAppRoutes extends AppRoutes {
@@ -10,10 +12,27 @@ class ExampleAppRoutes extends AppRoutes {
   @override
   Map<String, RouteConfig> get routes => {
     mainScreen: RouteConfig(
-      widgetRegister: (queryParams, fromSource) => const MainScreen(),
+      widgetRegister: (queryParams) => const MainScreen(),
     ),
     detailScreen: RouteConfig(
-      widgetRegister: (queryParams, fromSource) => const DetailScreen(),
+      widgetRegister: (queryParams) => const DetailScreen(),
+    ),
+    'sheet': RouteConfig(
+      actionRegister: (query) async {
+        await showDialog(
+          context: globalNavigatorKey.currentContext!,
+          builder: (context) => AlertDialog(
+            title: const Text('Deep Link Detected'),
+            content: Text(query['label'] ?? ''),
+            actions: [
+              TextButton(
+                onPressed: () => Navigator.of(context).pop(),
+                child: const Text('OK'),
+              ),
+            ],
+          ),
+        );
+      },
     ),
   };
 }