+++ example/lib/main.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_link_nav/flutter_link_nav.dart';
 import 'app_routes.dart';
 
+final globalNavigatorKey = GlobalObjectKey<NavigatorState>('nav');
 void main() {
   ExampleAppRoutes().registerRoutes();
   runApp(const MyApp());
@@ -13,6 +14,7 @@ class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
+      navigatorKey: globalNavigatorKey,
       initialRoute: ExampleAppRoutes.mainScreen,
       onGenerateRoute: AppRoutes.generateRoute,
     );