+++ lib/src/src.dart
@@ -1,3 +1,4 @@
 export 'route/route_registry.dart';
 export 'route/app_routes.dart';
+export 'utils/utils.dart';
 export 'deeplink/deeplink_handle.dart';